Neredeyse Bitti
Kodun çoğunu yazdım geriye test etmek kaldı.

+++ Minibüs.java
@@ -2,8 +2,7 @@ class Minibüs extends Araç {
 
     Minibüs(String HGSNumarası, String sürücüAdı, String sürücüSoyadı, int bakiye) {
         super(HGSNumarası, sürücüAdı, sürücüSoyadı, bakiye);
-
-        setAraçSınıfı((byte) 2);
+        setAraçSınıfı(2);
     }
