+++ Otobüs.java
@@ -1,9 +1,8 @@
 class Otobüs extends Araç{
 
-    Otobüs(String HGSNumarası, String sürücüAdı, String sürücüSoyadı, int bakiye) {
+    Otobüs(String HGSNumarası, String sürücüAdı, String sürücüSoyadı, int bakiye){
         super(HGSNumarası, sürücüAdı, sürücüSoyadı, bakiye);
-
-        setAraçSınıfı((byte) 3);
+        setAraçSınıfı(3);
     }
 
 }
