+++ Otomobil.java
@@ -2,7 +2,6 @@ class Otomobil extends Araç{
 
     Otomobil(String HGSNumarası, String sürücüAdı, String sürücüSoyadı, int bakiye) {
         super(HGSNumarası, sürücüAdı, sürücüSoyadı, bakiye);
-
-        setAraçSınıfı((byte) 1);
+        setAraçSınıfı(1);
     }
 }
